backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.6.0
pymongo==4.9.0
beanie==1.24.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
PyMuPDF==1.23.21
Pillow==10.2.0
opencv-python-headless==4.9.0.80
pytesseract==0.3.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
aiofiles==23.2.1
celery==5.3.4
redis==5.0.1
python-dateutil==2.8.2
numpy==1.26.3